+++ ADAs/Greene/ADA4/No visual/recorrerMatriz1-2Descendente.c
@@ -1,7 +1,7 @@
 /**
- * @file recorrerMatriz2-3Descendente.c
+ * @file recorrerMatriz1-2Descendente.c
  * @author Carlos Enrique Greene Mex (greenecarlos17@gmail.com)
- * @brief Programa que recorre la zona 2 a 3 forma descendente de una matriz de NxN
+ * @brief Programa que recorre la zona 1 a 2 forma descendente y no visual de una matriz de NxN
  * @version 0.1
  * @date 2020-03-22
  * 
@@ -21,7 +21,7 @@
  * @param arr 
  */
 void asignarArreglo(int arr[][N]);
-void recorrerMatriz2_3Descendente(int arr[][N]);
+void recorrerMatriz1_2Descendente(int arr[][N]);
 
 /**
  * @brief Main donde inicia el programa
@@ -36,7 +36,7 @@ int main()
 
     asignarArreglo(arreglo);
 
-    recorrerMatriz2_3Descendente(arreglo);
+    recorrerMatriz1_2Descendente(arreglo);
 
     return 0;
 }
@@ -58,18 +58,17 @@ void asignarArreglo(int arr[][N])
 }
 
 /**
- * @brief Función que imprime la zona 2 a 3 de forma descendente de la matriz
+ * @brief Función que imprime la zona 1 a 2 de forma descendente y no visual de la matriz
  * 
  * @param arr 
  */
-void recorrerMatriz2_3Descendente(int arr[][N])
+void recorrerMatriz1_2Descendente(int arr[][N])
 {
     for(int i=N-1; i>=0; i--)
     {
-        for (int j = i; j<N; j++)
+        for (int j = 0; j<(N-i); j++)
         {
             printf("%d,%d: %d\t", i, j, arr[i][j]);
         }
-        printf("\n");
     }
 }
