+++ ADAs/Greene/ADA4/No visual/recorrerMatriz3-4Descendente.c
@@ -1,7 +1,7 @@
 /**
- * @file recorrerMatriz2Ascendente.c
+ * @file recorrerMatriz3-4Ascendente.c
  * @author Carlos Enrique Greene Mex (greenecarlos17@gmail.com)
- * @brief Programa que recorre la zona 4 de forma descendente de una matriz de NxN
+ * @brief Programa que recorre la zona 3 a 4 de forma descendente y no visual de una matriz de NxN
  * @version 0.1
  * @date 2020-03-22
  * 
@@ -21,7 +21,7 @@
  * @param arr 
  */
 void asignarArreglo(int arr[][N]);
-void recorrerMatriz4Descendente(int arr[][N]);
+void recorrerMatriz3_4Descendente(int arr[][N]);
 
 /**
  * @brief Main donde inicia el programa
@@ -36,7 +36,7 @@ int main()
 
     asignarArreglo(arreglo);
 
-    recorrerMatriz4Descendente(arreglo);
+    recorrerMatriz3_4Descendente(arreglo);
 
     return 0;
 }
@@ -58,20 +58,17 @@ void asignarArreglo(int arr[][N])
 }
 
 /**
- * @brief Función que imprime la zona 4 forma de descendente de la matriz
+ * @brief Función que imprime la zona 3 a 4 de forma descendente y no visual de la matriz
  * 
  * @param arr 
  */
-void recorrerMatriz4Descendente(int arr[][N])
+void recorrerMatriz3_4Descendente(int arr[][N])
 {
     for(int i=N-1; i>=0; i--)
     {
-        for (int j=N-i-1; j<N; j++)
+        for (int j = (N-i-1); j<N; j++)
         {
-            if (i>=j)
-            {
-                printf("%d,%d: %d\t", i, j, arr[i][j]);
-            }
+            printf("%d,%d: %d\t", i, j, arr[i][j]);
         }
         printf("\n");
     }
