+++ ADAs/Greene/ADA4/No visual/recorrerMatriz3Ascendente.c
@@ -1,7 +1,7 @@
 /**
- * @file recorrerMatriz2-3Ascendente.c
+ * @file recorrerMatriz2Ascendente.c
  * @author Carlos Enrique Greene Mex (greenecarlos17@gmail.com)
- * @brief Programa que recorre la zona 2 a 3 forma ascendente de una matriz de NxN
+ * @brief Programa que recorre la zona 3 de forma ascendente y no visual de una matriz de NxN
  * @version 0.1
  * @date 2020-03-22
  * 
@@ -21,7 +21,7 @@
  * @param arr 
  */
 void asignarArreglo(int arr[][N]);
-void recorrerMatriz2_3Ascendente(int arr[][N]);
+void recorrerMatriz3Ascendente(int arr[][N]);
 
 /**
  * @brief Main donde inicia el programa
@@ -36,7 +36,7 @@ int main()
 
     asignarArreglo(arreglo);
 
-    recorrerMatriz2_3Ascendente(arreglo);
+    recorrerMatriz3Ascendente(arreglo);
 
     return 0;
 }
@@ -58,17 +58,22 @@ void asignarArreglo(int arr[][N])
 }
 
 /**
- * @brief Función que imprime la zona 2 a 3 de forma ascendente de la matriz
+ * @brief Función que imprime la zona 3 forma de ascendente y no visual de la matriz
  * 
  * @param arr 
  */
-void recorrerMatriz2_3Ascendente(int arr[][N])
+void recorrerMatriz3Ascendente(int arr[][N])
 {
+    int aux;
     for(int i=0; i<N; i++)
     {
-        for (int j = i; j<N; j++)
+        for (int j=0; j<N; j++)
         {
-            printf("%d,%d: %d\t", i, j, arr[i][j]);
+            aux=i+j;
+            if (i<=j && aux>=N-1)
+            {
+                printf("%d,%d: %d\t", i, j, arr[i][j]);
+            }
         }
         printf("\n");
     }
