+++ ADAs/Greene/ADA4/Visual/recorrerMatriz1Ascendente.c
@@ -1,7 +1,7 @@
 /**
- * @file recorrerMatriz2Descendente.c
+ * @file recorrerMatriz1Ascendente.c
  * @author Carlos Enrique Greene Mex (greenecarlos17@gmail.com)
- * @brief Programa que recorre la zona 2 de forma descendente de una matriz de NxN
+ * @brief Programa que recorre la zona 1 forma ascendente y visual de una matriz de NxN
  * @version 0.1
  * @date 2020-03-22
  * 
@@ -21,7 +21,7 @@
  * @param arr 
  */
 void asignarArreglo(int arr[][N]);
-void recorrerMatriz2Descendente(int arr[][N]);
+void recorrerMatriz1Ascendente(int arr[][N]);
 
 /**
  * @brief Main donde inicia el programa
@@ -36,7 +36,7 @@ int main()
 
     asignarArreglo(arreglo);
 
-    recorrerMatriz2Descendente(arreglo);
+    recorrerMatriz1Ascendente(arreglo);
 
     return 0;
 }
@@ -58,17 +58,17 @@ void asignarArreglo(int arr[][N])
 }
 
 /**
- * @brief Función que imprime la zona 2 de forma descendente de la matriz
+ * @brief Función que imprime la zona 1 forma ascendente y visual de la matriz
  * 
  * @param arr 
  */
-void recorrerMatriz2Descendente(int arr[][N])
+void recorrerMatriz1Ascendente(int arr[][N])
 {
-    for(int i=N-1; i>=0; i--)
+    for(int i=0; i<N; i++)
     {
         for (int j=0; j<=(N-i-1); j++)
         {
-            if (i<=j)
+            if (i>=j)
             {
                 printf("%d,%d: %d\t", i, j, arr[i][j]);
             }
