+++ ADAs/Greene/ADA4/Visual/recorrerMatriz1Descendente.c
@@ -1,7 +1,7 @@
 /**
- * @file recorrerMatriz2Ascendente.c
+ * @file recorrerMatriz1Descendente.c
  * @author Carlos Enrique Greene Mex (greenecarlos17@gmail.com)
- * @brief Programa que recorre la zona 3 de forma ascendente de una matriz de NxN
+ * @brief Programa que recorre la zona 1 forma descendente y visual de una matriz de NxN
  * @version 0.1
  * @date 2020-03-22
  * 
@@ -21,7 +21,7 @@
  * @param arr 
  */
 void asignarArreglo(int arr[][N]);
-void recorrerMatriz3Ascendente(int arr[][N]);
+void recorrerMatriz1Descendente(int arr[][N]);
 
 /**
  * @brief Main donde inicia el programa
@@ -36,7 +36,7 @@ int main()
 
     asignarArreglo(arreglo);
 
-    recorrerMatriz3Ascendente(arreglo);
+    recorrerMatriz1Descendente(arreglo);
 
     return 0;
 }
@@ -58,17 +58,17 @@ void asignarArreglo(int arr[][N])
 }
 
 /**
- * @brief Función que imprime la zona 3 forma de ascendente de la matriz
+ * @brief Función que imprime la zona 1 forma descendente y visual de la matriz
  * 
  * @param arr 
  */
-void recorrerMatriz3Ascendente(int arr[][N])
+void recorrerMatriz1Descendente(int arr[][N])
 {
-    for(int i=0; i<N; i++)
+    for(int i=N-1; i>=0; i--)
     {
-        for (int j=N-i-1; j<N; j++)
+        for (int j=0; j<=(N-i-1); j++)
         {
-            if (i<=j)
+            if (i>=j)
             {
                 printf("%d,%d: %d\t", i, j, arr[i][j]);
             }
