+++ ADAs/Greene/ADA4/Visual/recorrerMatriz2Ascendente.c
@@ -1,7 +1,7 @@
 /**
- * @file recorrerMatriz1Ascendente.c
+ * @file recorrerMatriz2Ascendente.c
  * @author Carlos Enrique Greene Mex (greenecarlos17@gmail.com)
- * @brief Programa que recorre la zona 1 forma ascendente de una matriz de NxN
+ * @brief Programa que recorre la zona 2 de forma ascendente y visual de una matriz de NxN
  * @version 0.1
  * @date 2020-03-22
  * 
@@ -21,7 +21,7 @@
  * @param arr 
  */
 void asignarArreglo(int arr[][N]);
-void recorrerMatriz1Ascendente(int arr[][N]);
+void recorrerMatriz2Ascendente(int arr[][N]);
 
 /**
  * @brief Main donde inicia el programa
@@ -36,7 +36,7 @@ int main()
 
     asignarArreglo(arreglo);
 
-    recorrerMatriz1Ascendente(arreglo);
+    recorrerMatriz2Ascendente(arreglo);
 
     return 0;
 }
@@ -58,20 +58,25 @@ void asignarArreglo(int arr[][N])
 }
 
 /**
- * @brief Función que imprime la zona 1 forma ascendente de la matriz
+ * @brief Función que imprime la zona 2 forma de ascendente y visual de la matriz
  * 
  * @param arr 
  */
-void recorrerMatriz1Ascendente(int arr[][N])
+void recorrerMatriz2Ascendente(int arr[][N])
 {
     for(int i=0; i<N; i++)
     {
         for (int j=0; j<=(N-i-1); j++)
         {
-            if (i>=j)
+            if (i<=j)
             {
                 printf("%d,%d: %d\t", i, j, arr[i][j]);
             }
+            else
+            {
+                printf("\t");
+            }
+            
         }
         printf("\n");
     }
