+++ ADAs/Greene/ADA4/Visual/recorrerMatriz3Ascendente.c
@@ -1,7 +1,7 @@
 /**
- * @file recorrerMatriz1-2Descendente.c
+ * @file recorrerMatriz2Ascendente.c
  * @author Carlos Enrique Greene Mex (greenecarlos17@gmail.com)
- * @brief Programa que recorre la zona 1 a 2 forma descendente de una matriz de NxN
+ * @brief Programa que recorre la zona 3 de forma ascendente y visual de una matriz de NxN
  * @version 0.1
  * @date 2020-03-22
  * 
@@ -13,7 +13,7 @@
 #include <time.h>
 
 //Se define el valor de N como constante
-#define N 3     
+#define N 5     
 
 /**
  * @brief Prototipos de las funciones
@@ -21,7 +21,7 @@
  * @param arr 
  */
 void asignarArreglo(int arr[][N]);
-void recorrerMatriz1_2Descendente(int arr[][N]);
+void recorrerMatriz3Ascendente(int arr[][N]);
 
 /**
  * @brief Main donde inicia el programa
@@ -36,7 +36,7 @@ int main()
 
     asignarArreglo(arreglo);
 
-    recorrerMatriz1_2Descendente(arreglo);
+    recorrerMatriz3Ascendente(arreglo);
 
     return 0;
 }
@@ -58,18 +58,26 @@ void asignarArreglo(int arr[][N])
 }
 
 /**
- * @brief Función que imprime la zona 1 a 2 de forma descendente de la matriz
+ * @brief Función que imprime la zona 3 forma de ascendente y visual de la matriz
  * 
  * @param arr 
  */
-void recorrerMatriz1_2Descendente(int arr[][N])
+void recorrerMatriz3Ascendente(int arr[][N])
 {
-    for(int i=N-1; i>=0; i--)
+    int aux;
+    for(int i=0; i<N; i++)
     {
-        for (int j = 0; j<(N-i); j++)
+        for (int j=0; j<N; j++)
         {
-            printf("%d,%d: %d\t", i, j, arr[i][j]);
+            aux=i+j;
+            if (i<=j && aux>=N-1)
+            {
+                printf("%d,%d: %d\t", i, j, arr[i][j]);
+            }
+            else if(j>=(N-1)/2)
+            {    
+                printf("\t");
+            }
         }
         printf("\n");
     }
-}
