+++ ADAs/Greene/ADA4/Visual/recorrerMatriz3Descendente.c
@@ -1,7 +1,7 @@
 /**
- * @file recorrerMatriz3-4Ascendente.c
+ * @file recorrerMatriz2Ascendente.c
  * @author Carlos Enrique Greene Mex (greenecarlos17@gmail.com)
- * @brief Programa que recorre la zona 3 a 4 de forma descendente de una matriz de NxN
+ * @brief Programa que recorre la zona 3 de forma descendente y visual de una matriz de NxN
  * @version 0.1
  * @date 2020-03-22
  * 
@@ -21,7 +21,7 @@
  * @param arr 
  */
 void asignarArreglo(int arr[][N]);
-void recorrerMatriz3_4Descendente(int arr[][N]);
+void recorrerMatriz3Descendente(int arr[][N]);
 
 /**
  * @brief Main donde inicia el programa
@@ -36,7 +36,7 @@ int main()
 
     asignarArreglo(arreglo);
 
-    recorrerMatriz3_4Descendente(arreglo);
+    recorrerMatriz3Descendente(arreglo);
 
     return 0;
 }
@@ -58,17 +58,26 @@ void asignarArreglo(int arr[][N])
 }
 
 /**
- * @brief Función que imprime la zona 3 a 4 de forma descendente de la matriz
+ * @brief Función que imprime la zona 3 forma de descendente y visual de la matriz
  * 
  * @param arr 
  */
-void recorrerMatriz3_4Descendente(int arr[][N])
+void recorrerMatriz3Descendente(int arr[][N])
 {
+    int aux;
     for(int i=N-1; i>=0; i--)
     {
-        for (int j = (N-i-1); j<N; j++)
+        for (int j=0; j<N; j++)
         {
-            printf("%d,%d: %d\t", i, j, arr[i][j]);
+            aux=i+j;
+            if (i<=j && aux>=N-1)
+            {
+                printf("%d,%d: %d\t", i, j, arr[i][j]);
+            }
+            else if(j>=(N-1)/2)
+            {    
+                printf("\t");
+            }
         }
         printf("\n");
     }
