+++ ADAs/Greene/ADA4/Visual/recorrerMatriz4Descendente.c
@@ -1,7 +1,7 @@
 /**
- * @file recorrerMatriz1Descendente.c
+ * @file recorrerMatriz2Ascendente.c
  * @author Carlos Enrique Greene Mex (greenecarlos17@gmail.com)
- * @brief Programa que recorre la zona 1 forma descendente de una matriz de NxN
+ * @brief Programa que recorre la zona 4 de forma descendente y visual de una matriz de NxN
  * @version 0.1
  * @date 2020-03-22
  * 
@@ -21,7 +21,7 @@
  * @param arr 
  */
 void asignarArreglo(int arr[][N]);
-void recorrerMatriz1Descendente(int arr[][N]);
+void recorrerMatriz4Descendente(int arr[][N]);
 
 /**
  * @brief Main donde inicia el programa
@@ -36,7 +36,7 @@ int main()
 
     asignarArreglo(arreglo);
 
-    recorrerMatriz1Descendente(arreglo);
+    recorrerMatriz4Descendente(arreglo);
 
     return 0;
 }
@@ -58,19 +58,25 @@ void asignarArreglo(int arr[][N])
 }
 
 /**
- * @brief Función que imprime la zona 1 forma descendente de la matriz
+ * @brief Función que imprime la zona 4 forma de descendente y visual de la matriz
  * 
  * @param arr 
  */
-void recorrerMatriz1Descendente(int arr[][N])
+void recorrerMatriz4Descendente(int arr[][N])
 {
+    int aux;
     for(int i=N-1; i>=0; i--)
     {
-        for (int j=0; j<=(N-i-1); j++)
+        for (int j=0; j<N; j++)
         {
-            if (i>=j)
+            aux=i+j;
+            if (i>=j && aux>=N-1)
             {
                 printf("%d,%d: %d\t", i, j, arr[i][j]);
+            }
+            else if(i>=(N-1)/2)
+            {
+                printf("\t");
             }
         }
         printf("\n");
